Cleaned up the repo
Each algorithm/DS has a folder which contains their explanation, java and/or python codes.
