+++ Pattern_searching KMP(Knuth Morris Pratt)_algo/KmpAlgo.py
