+++ Segment_Tree_For_sum_(Java & Python)/segment_Tree.py
